CHORE: subclass rather than pass update method
